django/hacker-news-api/requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.4
django-cache-url==3.4.0
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
environs==9.5.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.3.7
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
prompt-toolkit==3.0.29
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
